Search dialog: add searching by part number/msg code
The new "Components part number" checkbox is added to search dialog.
It enables searching the "msgcode" string attached to each element.
It is independent to existed checkbox - it was renamed to "Components name"

For me it is useful for cases like answering
"Where did they put that P-mosfet in this revision of design?"
(datasheets I'm working with has a lot of msgcode data - part number, Ohm values, even packages).

It's checkbox disabled by default,
since it's UI is a bit less expected then other searches -
the results that are showing below search fields are still names,
not msgcodes, so it is not so immediately obvious
why they are included there, also clicking them can lead to "not found"
if "Components name" is not selected too.

Also, I'm a not 100% sure of portability
of pointer-to-member-of-template class that I used in Searcher.cpp,
but I do not know actual problems with it.